+fixed build issue with wrong read/write/whatsnot in i386 builds +fixed debug output
git-svn-id: https://svn.code.sf.net/p/remotefs/code/trunk@890 7764f161-3e89-4957-9ae6-8e1f4830ae4a

+++ src/acl_utils_nss.c
@@ -85,7 +85,7 @@ char* patch_acl_for_server(const char *acl_text, const struct rfs_instance *inst
 	{
 		size_t last_len = strlen(start);
 
-		DEBUG("done: %ld, left: %ld, expected: %ld\n", done, last_len, patched_size);
+		DEBUG("done: %lu, left: %lu, expected: %lu\n", (long unsigned)done, (long unsigned)last_len, (long unsigned)patched_size);
 
 		if (done + last_len >= patched_size)
 		{
@@ -99,7 +99,7 @@ char* patch_acl_for_server(const char *acl_text, const struct rfs_instance *inst
 	
 	patched[done] = 0;
 
-	DEBUG("done patched: %ld, expected: %ld\n", done, patched_size - 1);
+	DEBUG("done patched: %lu, expected: %lu\n", (long unsigned)done, (long unsigned)patched_size - 1);
 
 	if (done < patched_size - 1)
 	{
